python-requirements-dev.txt
beautifulsoup4==4.14.2
cloudpickle==3.1.1
dateparser==1.2.2
docutils==0.22.2
lxml==6.0.2
nltk==3.9.2
gensim==4.3.3
joblib==1.5.2
elasticsearch==9.1.1
ipdb==0.13.13
numpy==1.26.4
memory-profiler==0.61.0
nose==1.3.7
num2words==0.5.14
pandas==2.3.3
psutil==7.1.0
pycountry==24.6.1
pytest-cov==7.0.0
pytest-pep8==1.0.6
pytest-pylint==0.21.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
regex==2025.9.18
reporters-db==3.2.58
requests==2.32.5
scikit-learn==1.3.2
scipy==1.13.1
sphinx==8.2.3
sphinx-rtd-theme==3.0.2
tika==3.1.0
twine==6.2.0
Unidecode==1.4.0
us==3.2.0
zahlwort2num==0.4.2
